[Feature #2] Refactor person class file
- remove unneccesary files and place related person classes into one file
